NGPF supported
This commit inculdes the code for creating a NGPF DeviceGroup and some
basic configuration function of ip and bgp. NGPF is quite a new function
in IxNetwork, the api has new pattern as well, which is MultipleValue
object. This commit is trying to give a simple way to deal with
MultipleValue object.